WIP: Create an image picker component
Created an image picker component with some of the functionality but not all. The functionalities I added was the ability for a user to click on the change button in the image upload component and the image picker component appears. Also, user can select any of the pre-populated images/posters and it will reflect in the image upload component but the images/posters are not persisted, so as soon as the user refreshes, it will reset to the defaul image/poster