WIP: Create an image picker component (#27)
* WIP: Create an image picker component

Created an image picker component with some of the functionality but not all. The functionalities I added was the ability for a user to click on the change button in the image upload component and the image picker component appears. Also, user can select any of the pre-populated images/posters and it will reflect in the image upload component but the images/posters are not persisted, so as soon as the user refreshes, it will reset to the defaul image/poster

* Fix deployment error

Commented out CLdImage and DefaultEventImage in EventImage.tsx in molecules folder to fix netlify deployment errors

* Set chip filter for image picker

* WIP: add chip filter functionality

* Add new model to hold images in db and new route to fetch it

Added a new model called ImageType to the database and created an API route to fetch it, so it can be displayed in the image picker

* Fix deployment error

* Fixing some UI

---------

Co-authored-by: muraliSingh7 <singhmurali.14@gmail.com>